ZDIOUT1: Add PRNDD to print the DD definition of a file
The global ^DD(FILE) is a file whose entries define the fields of FILE.
Fields of the field-definition entries are themselves defined in ^DD(0).
Add public entry point PRNDD to invoke the FILE entry point with proper
arguments to print the file definition instead of the file.  Add a FGR
argument to $$FILENAME() to use the name from the file header for DD
definition files.